start root reducer outside of Angular
