+++ modules/atlas/services/routing/state-url-converter.run.js
@@ -1,0 +1,13 @@
+(function () {
+    'use strict';
+
+    angular
+        .module('atlas')
+        .run(runBlock);
+
+    runBlock.$inject = ['$window', 'stateUrlConverter'];
+
+    function runBlock ($window, stateUrlConverter) {
+        $window.StateUrlConverter = stateUrlConverter;
+    }
+})();
