+++ modules/shared/services/redux/application-state.factory.js
@@ -3,9 +3,8 @@
         .module('dpShared')
         .factory('applicationState', applicationStateFactory);
 
-    applicationStateFactory.$inject = ['$window', 'Redux'];
-
-    function applicationStateFactory ($window, Redux) {
+    applicationStateFactory.$inject = ['$window', 'Redux', '$timeout', '$rootScope'];
+    function applicationStateFactory ($window, Redux, $timeout, $rootScope) {
         let reducer,
             stateUrlConverter;
 
@@ -20,7 +19,11 @@
             reducer = _reducer_;
             stateUrlConverter = _stateUrlConverter_;
 
-            $window.initializeState(Redux, _reducer_, _stateUrlConverter_, defaultState, ...middleware);
+            // Check if Angular component need to react to state change by subscribing to store.
+            const store = $window.initializeState(Redux, _reducer_, _stateUrlConverter_, defaultState, ...middleware);
+            store.subscribe(() => {
+                $timeout(() => $rootScope.$digest());
+            });
         }
     }
 })();
