+++ src/index.js
@@ -51,19 +51,17 @@ import '../modules/atlas/services/piwik/piwik.factory';
 import '../modules/atlas/services/piwik/piwik.run';
 import '../modules/atlas/services/redux/middleware/context-middleware.factory';
 import '../modules/atlas/services/redux/middleware/state-to-url-middleware.factory';
-import '../modules/atlas/services/redux/deprecated-reducer.factory';
-import '../modules/atlas/services/redux/reducer.factory';
 import '../modules/atlas/services/redux/reducers/data-selection-reducers.factory';
 import '../modules/atlas/services/redux/reducers/filters-reducers.factory';
-import '../modules/atlas/services/redux/reducers/home-reducers.factory';
-import '../modules/atlas/services/redux/reducers/page-reducers.factory';
 import '../modules/atlas/services/redux/reducers/search-reducers.factory';
 import '../modules/atlas/services/redux/reducers/straatbeeld-reducers.factory';
 import '../modules/atlas/services/redux/reducers/url-reducers.factory';
+import '../modules/atlas/services/redux/reducers/url-reducers.run';
 import '../modules/atlas/services/redux/store.run';
 import '../modules/atlas/services/routing/state-to-url.factory';
 import '../modules/atlas/services/routing/state-url-conversion.factory';
 import '../modules/atlas/services/routing/state-url-converter.factory';
+import '../modules/atlas/services/routing/state-url-converter.run';
 import '../modules/atlas/services/routing/url-to-state.factory';
 import '../modules/atlas/services/routing/url-to-state.run';
 import '../modules/data-selection/data-selection.module';
