+++ src/test-index.js
@@ -42,12 +42,8 @@ import '../modules/atlas/services/piwik/piwik.factory';
 import '../modules/atlas/services/piwik/piwik.run';
 import '../modules/atlas/services/redux/middleware/context-middleware.factory';
 import '../modules/atlas/services/redux/middleware/state-to-url-middleware.factory';
-import '../modules/atlas/services/redux/deprecated-reducer.factory';
-import '../modules/atlas/services/redux/reducer.factory';
 import '../modules/atlas/services/redux/reducers/data-selection-reducers.factory';
 import '../modules/atlas/services/redux/reducers/filters-reducers.factory';
-import '../modules/atlas/services/redux/reducers/home-reducers.factory';
-import '../modules/atlas/services/redux/reducers/page-reducers.factory';
 import '../modules/atlas/services/redux/reducers/search-reducers.factory';
 import '../modules/atlas/services/redux/reducers/straatbeeld-reducers.factory';
 import '../modules/atlas/services/redux/reducers/url-reducers.factory';
@@ -383,8 +379,6 @@ import '../modules/atlas/services/redux/middleware/state-to-url-middleware.facto
 import '../modules/atlas/services/redux/deprecated-reducer.factory.test';
 import '../modules/atlas/services/redux/reducers/url-reducers.factory.test';
 import '../modules/atlas/services/redux/reducers/search-reducers.factory.test';
-import '../modules/atlas/services/redux/reducers/page-reducers.factory.test';
-import '../modules/atlas/services/redux/reducers/home-reducers.factory.test';
 import '../modules/atlas/services/redux/reducers/data-selection-reducers.factory.test';
 import '../modules/atlas/services/redux/reducers/straatbeeld-reducers.factory.test';
 import '../modules/atlas/services/redux/reducers/filters-reducers.factory.test';
